refactor(charts): extract describe into ChartIntrospectionMixin
